refactor: move test to src
