+++ .eslintrc.yml
@@ -1,3 +1,5 @@
 extends: 'eslint-config-cheminfo'
 parserOptions:
   sourceType: module
+env:
+  jest: true
