+++ package.json
@@ -10,7 +10,7 @@
   ],
   "scripts": {
     "compile": "rollup -c",
-    "eslint": "eslint src test __tests__",
+    "eslint": "eslint src test",
     "eslint-fix": "npm run eslint -- --fix",
     "prepare": "npm run compile",
     "test": "run-s testonly eslint",
