+++ src/__tests__/matrix/dynamicMethods.js
@@ -1,4 +1,4 @@
-import Matrix from '../../src';
+import Matrix from '../..';
 
 describe('Dynamic methods on matrices', () => {
