+++ src/__tests__/matrix/dynamicOperators.js
@@ -1,4 +1,4 @@
-import Matrix from '../../src';
+import Matrix from '../..';
 
 describe('Dynamic operators on matrices', () => {
